+++ src/poly/tests/api/interface_test.py
@@ -9,7 +9,9 @@ from unittest.mock import MagicMock, patch
 import requests
 
 from poly.handlers.interface import AgentStudioInterface
+from poly.handlers.protobuf.commands_pb2 import Command
 from poly.handlers.sdk import SourcererAPIError
+from poly.resources import Entity, Topic, Variable
 from poly.tests.testing_utils import make_mock_response
 
 
@@ -95,14 +97,14 @@ class WrapperErrorTranslation(unittest.TestCase):
 
     def test_pull_resources_translates_http_error(self):
         """pull_resources converts a requests.HTTPError into a ValueError."""
-        self.interface.sync_client.pull_resources.side_effect = requests.HTTPError("boom")
+        self.interface.sync_client.pull_projection.side_effect = requests.HTTPError("boom")
 
         with self.assertRaises(ValueError):
             self.interface.pull_resources()
 
     def test_pull_resources_translates_sourcerer_error(self):
         """pull_resources converts a SourcererAPIError into a ValueError."""
-        self.interface.sync_client.pull_resources.side_effect = SourcererAPIError("boom")
+        self.interface.sync_client.pull_projection.side_effect = SourcererAPIError("boom")
 
         with self.assertRaises(ValueError):
             self.interface.pull_resources()
@@ -118,7 +120,7 @@ class WrapperErrorTranslation(unittest.TestCase):
 class PullResourcesWithProjection(unittest.TestCase):
     """Tests for pull_resources when a projection is supplied directly."""
 
-    @patch("poly.handlers.interface.SyncClientHandler.load_resources_from_projection")
+    @patch("poly.resources.resource.load_resources_from_projection")
     def test_projection_json_skips_api_and_uses_loader(self, mock_loader):
         """Passing projection_json bypasses the sync client and loads locally."""
         interface = AgentStudioInterface()
@@ -131,7 +133,72 @@ class PullResourcesWithProjection(unittest.TestCase):
         self.assertEqual(resources, {"loaded": "resources"})
         self.assertIs(returned_projection, projection)
         mock_loader.assert_called_once_with(projection)
-        interface.sync_client.pull_resources.assert_not_called()
+        interface.sync_client.pull_projection.assert_not_called()
+
+
+def _build_interface():
+    """Build an AgentStudioInterface with a mocked sync_client."""
+    interface = AgentStudioInterface()
+    interface.sync_client = MagicMock()
+    interface.sync_client.sdk.create_metadata.return_value = Command().metadata
+    return interface
+
+
+class QueueResources(unittest.TestCase):
+    """Tests for AgentStudioInterface.queue_resources ordering and command shape."""
+
+    def setUp(self):
+        self.interface = _build_interface()
+
+    def test_creates_deletes_and_updates_produce_commands(self):
+        """New, updated, and deleted resources each produce a queued command."""
+        new = {Entity: {"ENT-1": Entity(resource_id="ENT-1", name="new", entity_type="free_text")}}
+        updated = {
+            Entity: {"ENT-2": Entity(resource_id="ENT-2", name="upd", entity_type="free_text")}
+        }
+        deleted = {
+            Entity: {"ENT-3": Entity(resource_id="ENT-3", name="old", entity_type="free_text")}
+        }
+
+        commands = self.interface.queue_resources(
+            deleted_resources=deleted, new_resources=new, updated_resources=updated
+        )
+
+        types = [c.type for c in commands]
+        self.assertEqual(types, ["entity_delete", "entity_create", "entity_update"])
+        self.assertEqual(self.interface.sync_client.sdk.add_command_to_queue.call_count, 3)
+
+    def test_priority_create_types_are_queued_first(self):
+        """Variables (a priority-create type) are created before non-priority types."""
+        new = {
+            Topic: {"TOPIC-1": Topic(
+                resource_id="TOPIC-1", name="t", actions="", content="c", example_queries=[]
+            )},
+            Variable: {"VAR-1": Variable(resource_id="VAR-1", name="balance")},
+        }
+
+        commands = self.interface.queue_resources(
+            deleted_resources={}, new_resources=new, updated_resources={}
+        )
+
+        types = [c.type for c in commands]
+        self.assertLess(types.index("variable_create"), types.index("create_topic"))
+
+    def test_priority_delete_types_are_queued_first(self):
+        """Variables (a priority-delete type) are deleted before non-priority types."""
+        deleted = {
+            Topic: {"TOPIC-1": Topic(
+                resource_id="TOPIC-1", name="t", actions="", content="c", example_queries=[]
+            )},
+            Variable: {"VAR-1": Variable(resource_id="VAR-1", name="balance")},
+        }
+
+        commands = self.interface.queue_resources(
+            deleted_resources=deleted, new_resources={}, updated_resources={}
+        )
+
+        types = [c.type for c in commands]
+        self.assertLess(types.index("variable_delete"), types.index("delete_topic"))
 
 
 if __name__ == "__main__":
